Comparison of predictive equations for resting metabolic rate in healthy nonobese and obese adults: a systematic review.
1. J Am Diet Assoc. 2005 May;105(5):775-89.

Comparison of predictive equations for resting metabolic rate in healthy nonobese
and obese adults: a systematic review.

Frankenfield D(1), Roth-Yousey L, Compher C.

Author information: 
(1)Department of Clinical Nutrition, Milton S. Hershey Medical Center, Hershey,
PA, USA.

Comment in
    J Am Diet Assoc. 2011 Nov;111(11):1670-3.

BACKGROUND: An assessment of energy needs is a necessary component in the
development and evaluation of a nutrition care plan. The metabolic rate can be
measured or estimated by equations, but estimation is by far the more common
method. However, predictive equations might generate errors large enough to
impact outcome. Therefore, a systematic review of the literature was undertaken
to document the accuracy of predictive equations preliminary to deciding on the
imperative to measure metabolic rate.
METHODS: As part of a larger project to determine the role of indirect
calorimetry in clinical practice, an evidence team identified published articles 
that examined the validity of various predictive equations for resting metabolic 
rate (RMR) in nonobese and obese people and also in individuals of various ethnic
and age groups. Articles were accepted based on defined criteria and abstracted
using evidence analysis tools developed by the American Dietetic Association.
Because these equations are applied by dietetics practitioners to individuals, a 
key inclusion criterion was research reports of individual data. The evidence was
systematically evaluated, and a conclusion statement and grade were developed.
RESULTS: Four prediction equations were identified as the most commonly used in
clinical practice (Harris-Benedict, Mifflin-St Jeor, Owen, and World Health
Organization/Food and Agriculture Organization/United Nations University
[WHO/FAO/UNU]). Of these equations, the Mifflin-St Jeor equation was the most
reliable, predicting RMR within 10% of measured in more nonobese and obese
individuals than any other equation, and it also had the narrowest error range.
No validation work concentrating on individual errors was found for the
WHO/FAO/UNU equation. Older adults and US-residing ethnic minorities were
underrepresented both in the development of predictive equations and in
validation studies.
CONCLUSIONS: The Mifflin-St Jeor equation is more likely than the other equations
tested to estimate RMR to within 10% of that measured, but noteworthy errors and 
limitations exist when it is applied to individuals and possibly when it is
generalized to certain age and ethnic groups. RMR estimation errors would be
eliminated by valid measurement of RMR with indirect calorimetry, using an
evidence-based protocol to minimize measurement error. The Expert Panel advises
clinical judgment regarding when to accept estimated RMR using predictive
equations in any given individual. Indirect calorimetry may be an important tool 
when, in the judgment of the clinician, the predictive methods fail an individual
in a clinically relevant way. For members of groups that are greatly
underrepresented by existing validation studies of predictive equations, a high
level of suspicion regarding the accuracy of the equations is warranted.

DOI: 10.1016/j.jada.2005.02.005 
PMID: 15883556  [Indexed for MEDLINE]